+++ lib/task.h
@@ -3,10 +3,10 @@
 
 char *task_curr(char *prj);
 char *task_prev(char *prj);
-int task_chk(char *id);
-int task_chklen(char *id);
+int task_is_valid_name(char *id);
+int task_is_valid_length(char *id);
 int task_add(char *prj, char *id);
-int task_ext(char *prj, char *id);
+int task_exist(char *prj, char *id);
 int task_del(char *prj, char *id);
 int task_iscurr(char *prj, char *id);
 int task_isprev(char *prj, char *id);